requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
Pillow==8.2.0
pycodestyle==2.7.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.2.4
toml==0.10.2
gunicorn
psycopg2
